bookmark button moved to offer-page from offer-header
