+++ src/components/review/review.tsx
@@ -19,7 +19,6 @@ type ReviewProps = {
 function Review(reviewProps: ReviewProps): JSX.Element {
 
   const { date, user: { name, avatarUrl }, comment, rating } = reviewProps;
-
   return (
     <li className='reviews__item'>
       <div className='reviews__user user'>
